pyproject.toml
[project]
name = "sit-apartment-checker"
version = "0.1.0"
description = "Simple project to check for 1-bedroom apartments in SIT Trondheim"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
 "twilio>=9.6.2",
]
